payment/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.5
requests==2.31.0
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.3
